+++ tests/e2e/tx_test.go
@@ -25,7 +25,7 @@ func (s *IntegrationTestSuite) executeTransactions() {
 // SendBankTransfer sends a bank transfer to check that the transactions are still working after the upgrade.
 func (s *IntegrationTestSuite) sendBankTransfer(ctx context.Context) {
 	receiver := "evmos1jcltmuhplrdcwp7stlr4hlhlhgd4htqh3a79sq"
-	sentCoins := sdk.Coins{sdk.NewInt64Coin(config.GetDenom(), 10000000000)}
+	sentCoins := sdk.Coins{sdk.NewInt64Coin(config.GetEVMCoinDenom(), 10000000000)}
 
 	balancePre, err := s.upgradeManager.GetBalance(ctx, s.upgradeParams.ChainID, receiver)
 	s.Require().NoError(err, "can't get balance of receiver account")