+++ x/evm/config/denom_config.go
@@ -9,23 +9,31 @@ package config
 import (
 	"fmt"
 
+	"cosmossdk.io/math"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
+// Decimals is a wrapper around uint32 to represent the decimal representation
+// of a Cosmos coin.
 type Decimals uint32
 
 const (
-	SixDecimals      Decimals = 6
+	// SixDecimals is the Decimals used for Cosmos coin with 6 decimals.
+	SixDecimals Decimals = 6
+	// EighteenDecimals is the Decimals used for Cosmos coin with 18 decimals.
 	EighteenDecimals Decimals = 18
 )
 
 // EvmCoinInfo struct holds the name and decimals of the EVM denom. The EVM denom
 // is the token used to pay fees in the EVM.
 type EvmCoinInfo struct {
-	denom    string
-	decimals Decimals
+	Denom    string
+	Decimals Decimals
 }
 
+// evmCoinInfo hold the information of the coin used in the EVM as gas token. It
+// can only be set via `EVMConfigurator` before starting the app.
 var evmCoinInfo EvmCoinInfo
 
 // setEVMCoinDecimals allows to define the decimals used in the representation
@@ -35,7 +43,7 @@ func setEVMCoinDecimals(d Decimals) {
 		panic(fmt.Errorf("invalid decimal value %d; the evm supports only 6 and 18 decimals", d))
 	}
 
-	evmCoinInfo.decimals = d
+	evmCoinInfo.Decimals = d
 }
 
 // setEVMCoinDenom allows to define the denom of the coin used in the EVM.
@@ -43,21 +51,41 @@ func setEVMCoinDenom(denom string) {
 	if err := sdk.ValidateDenom(denom); err != nil {
 		panic(err)
 	}
-	evmCoinInfo.denom = denom
+	evmCoinInfo.Denom = denom
 }
 
 // GetEVMCoinDecimals returns the decimals used in the representation of the EVM
 // coin.
 func GetEVMCoinDecimals() Decimals {
-	return evmCoinInfo.decimals
+	return evmCoinInfo.Decimals
 }
 
-func GetDenom() string {
-	return evmCoinInfo.denom
+// GetEVMCoinDenom returns the denom used for the EVM coin.
+func GetEVMCoinDenom() string {
+	return evmCoinInfo.Denom
 }
 
 // setEVMCoinInfo allows to define denom and decimals of the coin used in the EVM.
 func setEVMCoinInfo(evmdenom EvmCoinInfo) {
-	setEVMCoinDenom(evmdenom.denom)
-	setEVMCoinDecimals(evmdenom.decimals)
+	setEVMCoinDenom(evmdenom.Denom)
+	setEVMCoinDecimals(evmdenom.Decimals)
+}
+
+// ConversionFactor returns the conversion factor between the Decimals value and
+// the 18 decimals representation, i.e. `EighteenDecimals`.
+//
+// NOTE: This function does not check if the Decimal instance is valid or
+// not and by default returns the conversion factor of 1, i.e. from 18 decimals
+// to 18 decimals.
+func (d Decimals) ConversionFactor() math.Int {
+	if d == SixDecimals {
+		return math.NewInt(1e12)
+	}
+
+	return math.NewInt(1)
+}
+
+func SetEVMCoinTEST(evmdenom EvmCoinInfo) {
+	setEVMCoinDenom(evmdenom.Denom)
+	setEVMCoinDecimals(evmdenom.Decimals)
 }