+++ x/evm/keeper/state_transition_test.go
@@ -361,7 +361,7 @@ func (suite *KeeperTestSuite) TestRefundGas() {
 	// for refund to work
 	// NOTE: everything should happen within the same block for
 	// feecollector account to remain funded
-	baseDenom := config.GetDenom()
+	baseDenom := config.GetEVMCoinDenom()
 
 	coins := sdk.NewCoins(sdk.NewCoin(
 		baseDenom,